run Cloudfrontification when remote publishing
