+++ models/classes/import/class.TestImport.php
@@ -24,7 +24,6 @@ use oat\oatbox\PhpSerializable;
 use oat\oatbox\PhpSerializeStateless;
 use oat\tao\model\import\ImportHandlerHelperTrait;
 use oat\tao\model\import\TaskParameterProviderInterface;
-use oat\taoQtiTest\models\render\QtiPackageImportPreprocessing;
 use oat\taoQtiTest\models\event\QtiTestImportEvent;
 use Zend\ServiceManager\ServiceLocatorAwareInterface;
 
@@ -72,17 +71,10 @@ class taoQtiTest_models_classes_import_TestImport implements tao_models_classes_
         try {
             $uploadedFile = $this->fetchUploadedFile($form);
 
-            $qtiPackageImportPreprocessingService = $this->getQtiPackageImportPreprocessing();
-            $preprocessingReport = $qtiPackageImportPreprocessingService->run($uploadedFile);
-
             // The zip extraction is a long process that can exceed the 30s timeout
             helpers_TimeOutHelper::setTimeOutLimit(helpers_TimeOutHelper::LONG);
 
             $report = taoQtiTest_models_classes_QtiTestService::singleton()->importMultipleTests($class, $uploadedFile);
-
-            if ($preprocessingReport) {
-                $report->add($preprocessingReport);
-            }
 
             helpers_TimeOutHelper::reset();
 
@@ -96,13 +88,5 @@ class taoQtiTest_models_classes_import_TestImport implements tao_models_classes_
         } catch (Exception $e) {
             return common_report_Report::createFailure($e->getMessage());
         }
-    }
-
-    /**
-     * @return QtiPackageImportPreprocessing
-     */
-    private function getQtiPackageImportPreprocessing()
-    {
-        return $this->getServiceLocator()->get(QtiPackageImportPreprocessing::SERVICE_ID);
     }
 }
